+++ providers/go-feature-flag/pkg/provider.go
@@ -3,31 +3,32 @@ package gofeatureflag
 import (
 	"context"
 	"fmt"
+	"log/slog"
 	"time"
 
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/controller"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/api"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/evaluator"
 	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/hook"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/util"
-	"github.com/open-feature/go-sdk-contrib/providers/ofrep"
-	of "github.com/open-feature/go-sdk/openfeature"
+	controller "github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/manager"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/model"
+	"github.com/open-feature/go-sdk/openfeature"
 )
 
+var _ openfeature.FeatureProvider = &Provider{}
+var _ openfeature.Tracker = &Provider{}
+var _ openfeature.ContextAwareStateHandler = &Provider{}
+var _ openfeature.EventHandler = &Provider{}
+var eventChannelBuffer = 5
+
 const providerName = "GO Feature Flag"
-const cacheableMetadataKey = "gofeatureflag_cacheable"
 
 type Provider struct {
-	ofrepProvider        *ofrep.Provider
-	cache                *controller.Cache
-	dataCollectorManager controller.DataCollectorManager
-	options              ProviderOptions
-	status               of.State
-	hooks                []of.Hook
-	goffAPI              controller.GoFeatureFlagAPI
-	pollingInfo          struct {
-		ticker  *time.Ticker
-		channel chan bool
-	}
-	events chan of.Event
+	options          ProviderOptions
+	dataCollectorMgr controller.DataCollectorManager
+	eventStream      chan openfeature.Event
+	evaluator        evaluator.Evaluator
+	hooks            []openfeature.Hook
+	logger           *slog.Logger
 }
 
 // NewProvider allows you to create a GO Feature Flag provider without any context.
@@ -41,237 +42,160 @@ func NewProviderWithContext(ctx context.Context, options ProviderOptions) (*Prov
 	if err := options.Validation(); err != nil {
 		return nil, err
 	}
-	ofrepOptions := make([]ofrep.Option, 0)
-	if options.APIKey != "" {
-		ofrepOptions = append(ofrepOptions, ofrep.WithBearerToken(options.APIKey))
-	}
-	if options.HTTPClient != nil {
-		ofrepOptions = append(ofrepOptions, ofrep.WithClient(options.HTTPClient))
-	}
-	ofrepOptions = append(ofrepOptions, ofrep.WithHeaderProvider(func() (key string, value string) {
-		return controller.ContentTypeHeader, controller.ApplicationJson
-	}))
-	ofrepProvider := ofrep.NewProvider(options.Endpoint, ofrepOptions...)
-	cacheCtrl := controller.NewCache(options.FlagCacheSize, options.FlagCacheTTL, options.DisableCache)
-
-	// Adding metadata to the GO Feature Flag provider to be sent to the exporter
 	if options.ExporterMetadata == nil {
 		options.ExporterMetadata = make(map[string]any)
 	}
 	options.ExporterMetadata["provider"] = "go"
 	options.ExporterMetadata["openfeature"] = true
 
-	goffAPI := controller.NewGoFeatureFlagAPI(controller.GoFeatureFlagApiOptions{
-		Endpoint:         options.Endpoint,
-		HTTPClient:       options.HTTPClient,
-		APIKey:           options.APIKey,
-		ExporterMetadata: options.ExporterMetadata,
+	goffAPI := api.NewGoFeatureFlagAPI(api.GoFeatureFlagAPIOptions{
+		Endpoint:             options.Endpoint,
+		DataCollectorBaseURL: options.DataCollectorBaseURL,
+		HTTPClient:           options.HTTPClient,
+		APIKey:               options.APIKey,
+		Headers:              options.Headers,
+		ExporterMetadata:     options.ExporterMetadata,
 	})
-	dataCollectorManager := controller.NewDataCollectorManager(
-		goffAPI,
-		options.DataCollectorMaxEventStored,
-		options.DataFlushInterval,
-	)
-	return &Provider{
-		ofrepProvider:        ofrepProvider,
-		cache:                cacheCtrl,
-		dataCollectorManager: dataCollectorManager,
-		options:              options,
-		goffAPI:              goffAPI,
-		events:               make(chan of.Event, 5),
-		hooks:                []of.Hook{},
-	}, nil
+
+	if options.Logger == nil {
+		options.Logger = slog.Default()
+	}
+
+	var dcm controller.DataCollectorManager
+	if !options.DataCollectorDisabled {
+		dcm = controller.NewDataCollectorManager(*goffAPI, options.DataCollectorMaxEventStored, options.DataCollectorCollectInterval)
+	}
+
+	eventStream := make(chan openfeature.Event, eventChannelBuffer)
+	ev := selectEvaluator(options, goffAPI, eventStream)
+	p := &Provider{
+		options:          options,
+		dataCollectorMgr: dcm,
+		eventStream:      eventStream,
+		evaluator:        ev,
+		logger:           options.Logger,
+	}
+	p.hooks = buildHooks(options, &p.dataCollectorMgr)
+	return p, nil
 }
 
-func (p *Provider) Metadata() of.Metadata {
-	return of.Metadata{
+// EventChannel implements [openfeature.EventHandler].
+func (p *Provider) EventChannel() <-chan openfeature.Event {
+	return p.eventStream
+}
+
+// InitWithContext implements [openfeature.ContextAwareStateHandler].
+func (p *Provider) InitWithContext(ctx context.Context, evaluationContext openfeature.EvaluationContext) error {
+	if !p.options.DataCollectorDisabled {
+		p.dataCollectorMgr.Start()
+	}
+	return p.evaluator.Init(ctx)
+}
+
+// ShutdownWithContext implements [openfeature.ContextAwareStateHandler].
+func (p *Provider) ShutdownWithContext(ctx context.Context) error {
+	if !p.options.DataCollectorDisabled {
+		p.dataCollectorMgr.Stop(ctx)
+	}
+	return p.evaluator.Shutdown(ctx)
+}
+
+// BooleanEvaluation implements [openfeature.FeatureProvider].
+func (p *Provider) BooleanEvaluation(ctx context.Context, flag string, defaultValue bool, flatCtx openfeature.FlattenedContext) openfeature.BoolResolutionDetail {
+	return p.evaluator.BooleanEvaluation(ctx, flag, defaultValue, flatCtx)
+}
+
+// FloatEvaluation implements [openfeature.FeatureProvider].
+func (p *Provider) FloatEvaluation(ctx context.Context, flag string, defaultValue float64, flatCtx openfeature.FlattenedContext) openfeature.FloatResolutionDetail {
+	return p.evaluator.FloatEvaluation(ctx, flag, defaultValue, flatCtx)
+}
+
+// Hooks implements [openfeature.FeatureProvider].
+func (p *Provider) Hooks() []openfeature.Hook { return p.hooks }
+
+// IntEvaluation implements [openfeature.FeatureProvider].
+func (p *Provider) IntEvaluation(ctx context.Context, flag string, defaultValue int64, flatCtx openfeature.FlattenedContext) openfeature.IntResolutionDetail {
+	return p.evaluator.IntEvaluation(ctx, flag, defaultValue, flatCtx)
+}
+
+// Metadata implements [openfeature.FeatureProvider].
+func (p *Provider) Metadata() openfeature.Metadata {
+	return openfeature.Metadata{
 		Name: fmt.Sprintf("%s Provider", providerName),
 	}
 }
 
-func (p *Provider) BooleanEvaluation(ctx context.Context, flag string, defaultValue bool, evalCtx of.FlattenedContext) of.BoolResolutionDetail {
-	if err := util.ValidateTargetingKey(evalCtx); err != nil {
-		return of.BoolResolutionDetail{
-			Value:                    defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
-		}
-	}
-	if cacheValue, err := p.cache.GetBool(flag, evalCtx); err == nil && cacheValue != nil {
-		cacheValue.Reason = of.CachedReason
-		return *cacheValue
-	}
-	res := p.ofrepProvider.BooleanEvaluation(ctx, flag, defaultValue, evalCtx)
-	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
-		_ = p.cache.Set(flag, evalCtx, res)
-	}
-	return res
+// ObjectEvaluation implements [openfeature.FeatureProvider].
+func (p *Provider) ObjectEvaluation(ctx context.Context, flag string, defaultValue any, flatCtx openfeature.FlattenedContext) openfeature.InterfaceResolutionDetail {
+	return p.evaluator.ObjectEvaluation(ctx, flag, defaultValue, flatCtx)
 }
 
-func (p *Provider) StringEvaluation(ctx context.Context, flag string, defaultValue string, evalCtx of.FlattenedContext) of.StringResolutionDetail {
-	if err := util.ValidateTargetingKey(evalCtx); err != nil {
-		return of.StringResolutionDetail{
-			Value:                    defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
-		}
-	}
-	if cacheValue, err := p.cache.GetString(flag, evalCtx); err == nil && cacheValue != nil {
-		cacheValue.Reason = of.CachedReason
-		return *cacheValue
-	}
-	res := p.ofrepProvider.StringEvaluation(ctx, flag, defaultValue, evalCtx)
-	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
-		_ = p.cache.Set(flag, evalCtx, res)
-	}
-	return res
+// StringEvaluation implements [openfeature.FeatureProvider].
+func (p *Provider) StringEvaluation(ctx context.Context, flag string, defaultValue string, flatCtx openfeature.FlattenedContext) openfeature.StringResolutionDetail {
+	return p.evaluator.StringEvaluation(ctx, flag, defaultValue, flatCtx)
 }
 
-func (p *Provider) FloatEvaluation(ctx context.Context, flag string, defaultValue float64, evalCtx of.FlattenedContext) of.FloatResolutionDetail {
-	if err := util.ValidateTargetingKey(evalCtx); err != nil {
-		return of.FloatResolutionDetail{
-			Value:                    defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
-		}
+// Track implements [openfeature.Tracker].
+func (p *Provider) Track(ctx context.Context, trackingEventName string, evaluationContext openfeature.EvaluationContext, details openfeature.TrackingEventDetails) {
+	if p.options.DataCollectorDisabled {
+		p.logger.Warn("Data collector is disabled, skipping tracking event", "trackingEventName", trackingEventName)
+		return
 	}
-	if cacheValue, err := p.cache.GetFloat(flag, evalCtx); err == nil && cacheValue != nil {
-		cacheValue.Reason = of.CachedReason
-		return *cacheValue
+
+	contextKind := "user"
+	if isAnonymous, ok := evaluationContext.Attribute("anonymous").(bool); ok && isAnonymous {
+		contextKind = "anonymousUser"
 	}
-	res := p.ofrepProvider.FloatEvaluation(ctx, flag, defaultValue, evalCtx)
-	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
-		_ = p.cache.Set(flag, evalCtx, res)
+	userKey := evaluationContext.TargetingKey()
+	if userKey == "" {
+		userKey = "undefined"
 	}
-	return res
+
+	evalCtxMap := map[string]any{}
+	for k, v := range evaluationContext.Attributes() {
+		evalCtxMap[k] = v
+	}
+
+	event := model.TrackingEvent{
+		Kind:              "tracking",
+		ContextKind:       contextKind,
+		UserKey:           userKey,
+		CreationDate:      time.Now().Unix(),
+		Key:               trackingEventName,
+		EvaluationContext: evalCtxMap,
+		TrackingDetails:   details.Attributes(),
+	}
+	if err := p.dataCollectorMgr.AddEvent(event); err != nil {
+		p.logger.Error("Failed to add tracking event to data collector", "error", err)
+	}
 }
 
-func (p *Provider) IntEvaluation(ctx context.Context, flag string, defaultValue int64, evalCtx of.FlattenedContext) of.IntResolutionDetail {
-	if err := util.ValidateTargetingKey(evalCtx); err != nil {
-		return of.IntResolutionDetail{
-			Value:                    defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
-		}
+// selectEvaluator selects the evaluator based on the evaluation type
+func selectEvaluator(options ProviderOptions, goffAPI *api.GoFeatureFlagAPI, eventStream chan openfeature.Event) evaluator.Evaluator {
+	if options.EvaluationType == EvaluationTypeRemote {
+		return evaluator.NewRemoteEvaluator(options.Endpoint, options.HTTPClient, options.APIKey, options.Headers)
 	}
-	if cacheValue, err := p.cache.GetInt(flag, evalCtx); err == nil && cacheValue != nil {
-		cacheValue.Reason = of.CachedReason
-		return *cacheValue
-	}
-	res := p.ofrepProvider.IntEvaluation(ctx, flag, defaultValue, evalCtx)
-	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
-		_ = p.cache.Set(flag, evalCtx, res)
-	}
-	return res
+	return evaluator.NewInprocessEvaluator(options.FlagChangePollingInterval, goffAPI, eventStream)
 }
 
-func (p *Provider) ObjectEvaluation(ctx context.Context, flag string, defaultValue any, evalCtx of.FlattenedContext) of.InterfaceResolutionDetail {
-	if err := util.ValidateTargetingKey(evalCtx); err != nil {
-		return of.InterfaceResolutionDetail{
-			Value:                    defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
-		}
+// buildHooks constructs the list of hooks for the provider.
+func buildHooks(options ProviderOptions, dcm *controller.DataCollectorManager) []openfeature.Hook {
+	hooks := []openfeature.Hook{
+		hook.NewEvaluationEnrichmentHook(options.ExporterMetadata),
 	}
-	if cacheValue, err := p.cache.GetInterface(flag, evalCtx); err == nil && cacheValue != nil {
-		cacheValue.Reason = of.CachedReason
-		return *cacheValue
+	if options.EvaluationType != EvaluationTypeRemote && !options.DataCollectorDisabled {
+		hooks = append(hooks, hook.NewDataCollectorHook(dcm))
 	}
-	res := p.ofrepProvider.ObjectEvaluation(ctx, flag, defaultValue, evalCtx)
-	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
-		_ = p.cache.Set(flag, evalCtx, res)
-	}
-	return res
+	return hooks
 }
 
-func (p *Provider) Hooks() []of.Hook {
-	return p.hooks
-}
-
-// Init holds initialization logic of the provider
-func (p *Provider) Init(_ of.EvaluationContext) error {
-	p.hooks = append(p.hooks, hook.NewEvaluationEnrichmentHook(p.options.ExporterMetadata))
-	if !p.options.DisableDataCollector {
-		dataCollectorHook := hook.NewDataCollectorHook(&p.dataCollectorManager)
-		p.hooks = append(p.hooks, dataCollectorHook)
-		p.dataCollectorManager.Start()
-	}
-
-	// Start polling to check if there is any flag change in order to invalidate the cache.
-	if p.options.FlagChangePollingInterval >= 0 && !p.options.DisableCache {
-		p.startPolling(p.options.FlagChangePollingInterval)
-	}
-
-	p.status = of.ReadyState
-	p.events <- of.Event{
-		ProviderName: providerName, EventType: of.ProviderReady,
-		ProviderEventDetails: of.ProviderEventDetails{Message: "Provider is ready"}}
+// Init implements [openfeature.ContextAwareStateHandler].
+func (p *Provider) Init(evaluationContext openfeature.EvaluationContext) error {
+	// nothing to do here since we are using the context aware state handler
 	return nil
 }
 
-// Status exposes the status of the provider
-func (p *Provider) Status() of.State {
-	return p.status
-}
-
-// Shutdown defines the shutdown operation of the provider
+// Shutdown implements [openfeature.ContextAwareStateHandler].
 func (p *Provider) Shutdown() {
-	if !p.options.DisableDataCollector {
-		p.hooks = []of.Hook{}
-		p.dataCollectorManager.Stop()
-	}
-	p.stopPolling()
-}
-
-// EventChannel returns the event channel of this provider
-func (p *Provider) EventChannel() <-chan of.Event {
-	return p.events
-}
-
-// startPolling starts the polling mechanism that checks if the configuration has changed.
-func (p *Provider) startPolling(pollingInterval time.Duration) {
-	if pollingInterval == 0 {
-		pollingInterval = 120000 * time.Millisecond
-	}
-	p.pollingInfo = struct {
-		ticker  *time.Ticker
-		channel chan bool
-	}{
-		ticker:  time.NewTicker(pollingInterval),
-		channel: make(chan bool),
-	}
-	go func() {
-		for {
-			select {
-			case <-p.pollingInfo.channel:
-				return
-			case <-p.pollingInfo.ticker.C:
-				changeStatus, err := p.goffAPI.ConfigurationHasChanged()
-				switch changeStatus {
-				case controller.FlagConfigurationInitialized,
-					controller.FlagConfigurationNotChanged:
-					// do nothing
-
-				case controller.FlagConfigurationUpdated:
-					// Clearing the cache when the configuration is updated
-					p.cache.Purge()
-					p.events <- of.Event{
-						ProviderName: providerName, EventType: of.ProviderConfigChange,
-						ProviderEventDetails: of.ProviderEventDetails{Message: "Configuration has changed"}}
-				case controller.ErrorConfigurationChange:
-					p.events <- of.Event{
-						ProviderName: providerName, EventType: of.ProviderStale,
-						ProviderEventDetails: of.ProviderEventDetails{
-							Message: fmt.Sprintf("Impossible to check configuration change: %s", err),
-						},
-					}
-				}
-			}
-		}
-	}()
-}
-
-// stopPolling stops the polling mechanism that check if the configuration has changed.
-func (p *Provider) stopPolling() {
-	if p.pollingInfo.channel != nil {
-		p.pollingInfo.channel <- true
-	}
-	if p.pollingInfo.ticker != nil {
-		p.pollingInfo.ticker.Stop()
-	}
+	// nothing to do here since we are using the context aware state handler
 }